requirements.txt
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0
psycopg2-binary>=2.9.9
gunicorn>=21.2.0
python-dotenv>=1.0.0
django-cors-headers>=4.3.0
djangorestframework-simplejwt>=5.3.0
drf-spectacular>=0.26.5
prometheus-client>=0.17.1
django-prometheus>=2.3.1
pytest>=7.4.3
pytest-django>=4.7.0
pytest-cov>=4.1.0
black>=23.11.0
flake8>=6.1.0
sentry-sdk>=1.32.0
django-filter>=23.3
django-redis>=5.4.0